* script-base-FileMigrationLoader:
  Script based migration loading now works and all tests pass
  Tests are close
  Script based implementation. Poke-testing works.

+++ src/main/java/org/apache/ibatis/migration/operations/UpOperation.java
@@ -1,5 +1,5 @@
 /**
- *    Copyright 2010-2017 the original author or authors.
+ *    Copyright 2010-2018 the original author or authors.
  *
  *    Licensed under the Apache License, Version 2.0 (the "License");
  *    you may not use this file except in compliance with the License.
@@ -44,7 +44,7 @@ public final class UpOperation extends DatabaseOperation {
   public UpOperation(Integer steps) {
     super();
     this.steps = steps;
-    if (steps != null && steps.intValue() < 1) {
+    if (steps != null && steps < 1) {
       throw new IllegalArgumentException("step must be positive number or null.");
     }
   }
